requirements.txt
streamlit
pandas
gspread
google-auth
pyyaml
streamlit-authenticator
